+++ src/lib.rs
@@ -4,6 +4,7 @@ extern crate lazy_static;
 mod activities;
 mod etw;
 mod layer;
+mod providerwrapper;
 mod user_events;
 
 pub use layer::*;
